cleaning up loose prop ends
